+++ index.js
@@ -1,4 +1,4 @@
-import App from "./js/App.js";
+import App from "./src/App.js";
 
 const root = document.querySelector(".root");
 new App(root, "light");
