+++ src/App/Main/MainContent/NewsListView/Contents.js
