+++ src/App/Main/MainContent/NewsListView/Contents/Header.js
