+++ src/App/Main/MainContent/NewsListView/Contents/News.js
