+++ src/App/Main/MainContent/NewsListView/Contents/News/MainNews.js
