+++ src/App/RollingContents/RollingContents.js
@@ -3,6 +3,7 @@
 */
 const WAITING_TIME = 1000;
 const ANIMATION_TIME = 400;
+const MOVE_TIME = ANIMATION_TIME;
 const MOVE_PIXEL = -16;
 const ROLLING_INTERVAL = 5000;
 
@@ -26,7 +27,7 @@ const startRolling = function (rollingElement) {
     rollingElement.style.transitionDuration = `${ANIMATION_TIME}ms`;
     rollingElement.style.marginTop = `${MOVE_PIXEL}px`;
 
-    setTimeout(removeRollingMotion, ANIMATION_TIME);
+    setTimeout(removeRollingMotion, MOVE_TIME);
   };
 
   return setInterval(moveUpElement, ROLLING_INTERVAL);
